Harden Vercel deploy: drop embedded DB credentials, google-driver skip

+++ app/Providers/AppServiceProvider.php
@@ -22,7 +22,7 @@ class AppServiceProvider extends ServiceProvider
             URL::forceScheme('https');
         }
 
-        if (! $this->app->runningInConsole()) {
+        if (! $this->app->runningInConsole() && config('gsheet.driver') !== 'google') {
             DB::reconnect();
             DB::purge();
         }
